chore: move question edit modal files to new directory
