+++ apps/frontend-manage/src/components/questions/manipulation/ElementTypeMonitor.tsx
