+++ apps/frontend-manage/src/components/questions/manipulation/helpers.ts
