+++ apps/frontend-manage/src/pages/index.tsx
@@ -28,7 +28,7 @@ import { Suspense, useEffect, useMemo, useState } from 'react'
 import { isEmpty, pickBy } from 'remeda'
 import { buildIndex, processItems } from 'src/lib/utils/filters'
 import Layout from '../components/Layout'
-import QuestionEditModal from '../components/questions/QuestionEditModal'
+import QuestionEditModal from '../components/questions/manipulation/QuestionEditModal'
 import QuestionList from '../components/questions/QuestionList'
 import TagList from '../components/questions/tags/TagList'
 import ElementCreation, {
